No longer autocomplete all class names via nREPL.
Completing *all* 23k classes was crazy slow! Instead, cache them locally
and only complete them if ac-prefix contains a .

This means that you might have to write com. and then delete and re-add
the . to see the full class completion list. This feels acceptable as
it's not likely to be a common thing to do, and if you are doing it,
it's an explicit action, and therefore waiting for the list to generate
and appear is more acceptable.

Normal completions seem to be snappy again.